Revert "front-page add sidebar"
