+++ .gitignore
@@ -1,7 +1,3 @@
 node_modules
 package-lock.json
 out/
-res/
-.env
-src/static/icon/
-
