+++ add-sensor.html
@@ -1,0 +1,72 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="https://code.jquery.com/jquery-3.3.1.js"></script>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.3.0/css/all.min.css" />
+    <style>
+        body {
+            /* background-color: rgb(83, 83, 83); */
+            background-color: rgb(250, 250, 250);
+            height: 100vh;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+        }
+
+        .main-panel {
+            background-color: pink;
+            width: 480px;
+
+        }
+
+        #back-btn {
+            background: black;
+            border: 1px solid black;
+            border-radius: 4px;
+            padding: 4px;
+            margin-top: 8px;
+            float: right;
+            color: white;
+        }
+
+        #back-btn:hover {
+            background: white;
+            color: black;
+        }
+        #submit-btn {
+            margin-top: 10px;
+            float: right;
+        }
+
+        header {
+            display: box;
+        }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <form>
+            <header>
+                <h1 style="float: left;">Sensor Registration</h1>
+                <a href="./index.html"><i id="back-btn" class="fa fa-rotate-left"></i></a>
+            </header>
+            <br>
+            <br>
+            <br>
+            <small>Name</small>
+            <input type="text" class="form-control">
+            <small>Description</small>
+            <textarea class="form-control"></textarea>
+            
+            <button id="submit-btn" class="btn btn-primary">submit</button>
+        </form>
+    </div>
+    
+</body>
+</html>
