+++ index.html
@@ -1,0 +1,80 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="UTF-8">
+    <!-- <meta http-equiv="Content-Security-Policy"
+        content="default-src 'self'; script-src 'self'; style-src 'self' 'unsafe-inline'"> -->
+
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="https://code.jquery.com/jquery-3.3.1.js"></script>
+    <style>
+        body {
+            /* background-color: rgb(83, 83, 83); */
+            background-color: rgb(250, 250, 250);
+            height: 100vh;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+        }
+
+        .main-panel {
+            background-color: rgb(250, 250, 250);
+            border-radius: 10px;
+            padding: 16px;
+
+        }
+
+        .card {
+            margin-top: 16px;
+        }
+
+        .card-slot > .card:hover {
+            border: 1px solid black;
+            padding: -10px;
+        }
+
+        #add-btn {
+            background-color: rgb(59, 59, 59);
+            border: 0;
+        }
+    </style>
+    <script>
+        class Sensor {
+            constructor(name, desc) {
+                this.name = name
+                this.desc = desc
+            }
+
+            toDom() {
+                let card = document.createElement("div")
+                card.className = "card"
+                card.innerHTML = `
+                    <div class="card-body">
+                        <b>Sensor Name:</b> ${this.name} <br>
+                        <small>
+                            <b>Description: </b> ${this.desc} <b>Endpoint:</b> /test-sensor <b>Last Block: </b>EF8AC325
+                        </small>
+                    </div>`
+                return card
+            }
+        }
+        
+        window.onload = () => {
+            $(".card-slot").append(new Sensor("alice").toDom())
+            $(".card-slot").append(new Sensor("alice").toDom())
+        }
+    </script>
+</head>
+
+<body>
+    <br>
+    <div class="main-panel container">
+        <h1 style="float: left;">Sensors</h1>
+        <button id="add-btn" onclick="window.location.href='./add-sensor.html';" type="button" class="btn btn-primary" style="float: right;">Add sensor</button><br>
+        <br>
+        <div class="card-slot"></div>
+</body>
+
+</html>
