--- src/static/css/index.css
@@ -1,28 +1,0 @@
-body {
-    /* background-color: rgb(83, 83, 83); */
-    background-color: rgb(250, 250, 250);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.main-panel {
-    background-color: pink;
-    width: 480px;
-}
-
-.card {
-    margin-top: 16px;
-}
-
-.card-slot > .card:hover {
-    border: 1px solid black;
-    padding: -10px;
-}
-
-#add-btn {
-    background-color: rgb(59, 59, 59);
-    border: 0;
-}
-header {
-    display: box;
-}
