--- src/static/css/main.css
@@ -1,90 +1,0 @@
-.bd-placeholder-img {
-    font-size: 1.125rem;
-    text-anchor: middle;
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    user-select: none;
-  }
-
-  @media (min-width: 768px) {
-    .bd-placeholder-img-lg {
-      font-size: 3.5rem;
-    }
-  }
-
-  .b-example-divider {
-    height: 3rem;
-    background-color: rgba(0, 0, 0, .1);
-    border: solid rgba(0, 0, 0, .15);
-    border-width: 1px 0;
-    box-shadow: inset 0 .5em 1.5em rgba(0, 0, 0, .1), inset 0 .125em .5em rgba(0, 0, 0, .15);
-  }
-
-  .b-example-vr {
-    flex-shrink: 0;
-    width: 1.5rem;
-    height: 100vh;
-  }
-
-  .bi {
-    vertical-align: -.125em;
-    fill: currentColor;
-  }
-
-  .nav-scroller {
-    position: relative;
-    z-index: 2;
-    height: 2.75rem;
-    overflow-y: hidden;
-  }
-
-  .nav-scroller .nav {
-    display: flex;
-    flex-wrap: nowrap;
-    padding-bottom: 1rem;
-    margin-top: -1px;
-    overflow-x: auto;
-    text-align: center;
-    white-space: nowrap;
-    -webkit-overflow-scrolling: touch;
-  }
-
-
-
-
-
-body {
-    /* background-color: rgb(83, 83, 83); */
-    background-color: rgb(250, 250, 250);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.main-panel {
-    background-color: pink;
-    width: 480px;
-}
-#back-btn {
-    background: black;
-    border: 1px solid black;
-    border-radius: 4px;
-    padding: 4px;
-    margin-top: 8px;
-    float: right;
-    color: white;
-}
-#back-btn:hover {
-    background: white;
-    color: black;
-}
-#submit-btn {
-    margin-top: 10px;
-    float: right;
-}
-header {
-    display: box;
-}
-
-main html{
-    height: 100%;
-}
