--- src/static/css/sidebars.css
@@ -1,20 +1,0 @@
-.nav-link.active{
-    border-radius: 10px;
-    background-color: black;
-}
-.rounded-0{
-    width:90%;
-    margin-left: 5%;
-    border-radius:100%!important
-}
-.rounded-0.active{
-    width:90%;
-    margin-left: 5%;
-    border-radius:12%!important
-}
-.nav-pills .nav-link{
-    background:0 0;
-    border: 1px solid #dee2e6;
-    border-radius:var(--bs-nav-pills-border-radius);
-    border-radius:10px;
-}
