--- src/static/main.html
@@ -1,100 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-  <!-- bootstrap 코드 -->
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta name="description" content="">
-  <meta name="author" content="Mark Otto, Jacob Thornton, and Bootstrap contributors">
-  <meta name="generator" content="Hugo 0.108.0">
-  <title>Sidebars · Bootstrap v5.3</title>
-
-  <link rel="canonical" href="https://getbootstrap.com/docs/5.3/examples/sidebars/">
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet">
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
-  <script src="https://code.jquery.com/jquery-3.3.1.js"></script>
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.3.0/css/all.min.css" />
-
-  
-  <!-- bootstrap 끝 -->
-</head>
-
-<!-- SVG 아이콘 불러오기(아이콘은 외부에서 불러옴) -->
-<div id="external-content"></div>
-<script>
-  var xhr = new XMLHttpRequest();
-  xhr.open('GET', './icon/mainSvg.html', true);
-  xhr.onreadystatechange = function () {
-    if (xhr.readyState === 4 && xhr.status === 200) {
-      document.getElementById('external-content').innerHTML = xhr.responseText;
-    }
-  };
-  xhr.send();
-</script>
-<!-- SVG 아이콘 불러오기 끝 -->
-
-
-<!-- 사이드바 -->
-<div class="col-md-2">
-  <main class="d-flex flex-nowrap">
-    <!-- <div class="b-example-divider b-example-vr"></div> -->
-
-    <div class="d-flex flex-column flex-shrink-0 bg-light" style="width: 4.5rem;">
-      <a href="/" class="d-block p-3 link-dark text-decoration-none" title="Icon-only" data-bs-toggle="tooltip"
-        data-bs-placement="right">
-        <i class="fa-brands fa-html5 fa-3x" > </i>
-        <span class="visually-hidden">Icon-only</span>
-      </a>
-      <ul class="nav nav-pills nav-flush flex-column mb-auto text-center">
-        <li class="nav-item">
-          <a href="#" id="sensorView" class="nav-link active py-3 border-bottom rounded-0" aria-current="page"
-            title="Sensor" data-bs-toggle="tooltip"  data-bs-placement="right">
-            <i class="fa-solid fa-wifi"  role="img" aria-label="Sensor"></i>
-          </a>
-        </li>
-        <li>
-          <a href="#" id="searchView" class="nav-link py-3 border-bottom rounded-0" title="Search"
-            data-bs-toggle="tooltip" data-bs-placement="right">
-            <i class="fa-solid fa-magnifying-glass" aria-label="Search"></i>
-          </a>
-        </li>
-        <li>
-          <a href="#" id="settingView" class="nav-link py-3 border-bottom rounded-0" title="Setting"
-            data-bs-toggle="tooltip" data-bs-placement="right">
-            <i class="fa-solid fa-gear" role="img" aria-label="Setting"></i>
-
-          </a>
-        </li>
-        <li>
-          <div class="border-bottom py-3" style="height:100%; width:100%;">
-
-          </div>
-        </li>
-      </ul>
-    </div>
-  </main>
-</div>
-<!-- 사이드바 끝 -->
-
-<!-- 내용 시작 innerHTML 방식 -->
-<div id="contents" class="container col-md-10"> </div>
-
-
-</body>
-
-</html>
-
-<!--  사이드바 js 불러오기 -->
-<script src="./js/sidebar.js"></script>
-
-
-<!-- customs sidebar css -->
-<link href="./css/sidebars.css" rel="stylesheet">
-
-
-
-<!-- main css, js 블러우는 코드 -->
-<link rel="stylesheet" href="./css/main.css">
-<script src="./js/main.js"></script>
-
