--- src/static/sensorView-addSensor.html
@@ -1,15 +1,0 @@
-<form>
-    <header>
-        <h1 style="float: left;">Sensor Registration</h1>
-        <a href="./index.html"><i id="back-btn" class="fa fa-rotate-left"></i></a>
-    </header>
-    <br>
-    <br>
-    <br>
-    <small>Name</small>
-    <input type="text" class="form-control">
-    <small>Description</small>
-    <textarea class="form-control"></textarea>
-
-    <button id="submit-btn" class="btn btn-primary">submit</button>
-</form>
