--- src/static/sensorView.html
@@ -1,9 +1,0 @@
-<form>
-    <header>
-        <h1 style="float: left;">Sensor</h1>
-    </header>
-    <button id="add-btn" onclick="window.location.href='./add-sensor.html';" type="button" class="btn btn-primary"
-        style="float: right;">Add sensor</button>
-    <br>
-    <div class="card-slot"></div>
-</form>
